@@ -3,54 +3,54 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ag-grid-community/client-side-row-model": "27.2.1",
-    "@ag-grid-community/csv-export": "27.2.1",
-    "@ag-grid-community/react": "27.2.1",
+    "@ag-grid-community/all-modules": "^27.3.0",
+    "@ag-grid-community/client-side-row-model": "^29.0.0",
+    "@ag-grid-community/csv-export": "^29.0.0",
+    "@ag-grid-community/react": "^29.0.0",
     "@babel/runtime": "^7.17.8",
-    "@date-io/date-fns": "^2.13.2",
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/pickers": "^3.3.10",
-    "@mui/icons-material": "^5.6.0",
-    "@mui/material": "^5.0.0",
-    "@mui/x-data-grid": "^5.8.0",
-    "@mui/x-data-grid-generator": "^5.8.0",
-    "@mui/x-date-pickers": "^5.0.0-alpha.2",
+    "@date-io/date-fns": "^2.16.0",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@jridgewell/sourcemap-codec": "^1.4.14",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.11.4",
+    "@mui/x-data-grid": "^5.17.19",
+    "@mui/x-data-grid-generator": "^5.17.19",
+    "@mui/x-date-pickers": "^5.0.14",
     "@react-buddy/ide-toolbox": "^2.1.2",
     "@react-buddy/palette-mui": "^5.0.1",
+    "@rollup/plugin-terser": "^0.3.0",
     "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "ag-grid-community": "^27.2.1",
-    "ag-grid-react": "27.2.1",
-    "axios": "^0.26.1",
-    "axios-hooks": "^3.0.2",
+    "@testing-library/user-event": "^14.4.3",
+    "ag-grid-community": "^29.0.0",
+    "ag-grid-react": "^29.0.0",
+    "axios": "^1.2.2",
+    "axios-hooks": "^4.0.0",
     "bootstrap": "^5.1.3",
     "cors": "^2.8.5",
     "cross-env": "^7.0.3",
-    "date-fns": "^2.28.0",
+    "date-fns": "^2.29.3",
     "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-webpack-plugin": "^3.2.0",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.4.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.30.0",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "5.0.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.6.2",
+    "react-scripts": "^5.0.1",
     "redux": "^4.1.2",
     "redux-devtools-extension": "^2.13.9",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.4.1",
     "use": "^3.1.1",
     "validator": "^13.7.0",
-    "web-vitals": "^2.1.4",
-    "webpack": "^5.70.0",
+    "web-vitals": "^3.1.1",
     "webpack-dev-server": "^4.8.1"
   },
   "scripts": {
-    "start": "cross-env BROWSER=google-chrome-stable react-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
