@@ -1,7 +1,9 @@
 import React from 'react';
-import Routes  from './auth/Routes';
+import Paths  from './auth/Paths';
 import './App.css';
 import { NavbarSelect } from "./NavBar/NavBarSelect";
+
+
 
 const App = () =>  {
     // enable for a single run of app if debugger no longer stops at breakpionts
@@ -10,7 +12,7 @@ const App = () =>  {
         <React.StrictMode>
             <div>
                 <NavbarSelect/>
-                <Routes/>
+                <Paths/>
             </div>
         </React.StrictMode>
     )
